Created a basic home page. Added a basic nav bar to existing pages. Nav bar uses router which I installed with: npm install react-router-dom@6.22.0 Solving merge conflicts
