--- frontend/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "frontend",
-  "license": "0BSD",
-  "version": "1.0.0",
-  "main": "index.ts",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web"
-  },
-  "dependencies": {
-    "@expo/metro-runtime": "~4.0.1",
-    "expo": "~52.0.43",
-    "expo-status-bar": "~2.0.1",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-native": "0.76.9",
-    "react-native-web": "~0.19.13"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.25.2",
-    "@types/react": "~18.3.12",
-    "@types/react-native": "^0.73.0",
-    "typescript": "^5.8.3"
-  },
-  "private": true
-}
